sdk/python/flatagents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flatagents"
version = "0.10.0"
description = "A lightweight framework for building LLM-powered agents."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { email = "memgrafter@gmail.com" }
]
keywords = ["llm", "agents", "ai", "framework", "litellm", "aisuite"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = ["pyyaml", "jinja2", "aiofiles", "httpx"]

[project.optional-dependencies]
litellm = ["litellm"]
aisuite = ["aisuite[all]"]
validation = ["jsonschema>=4.0"]
metrics = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
orchestration = ["flatmachines>=0.10.0"]
# Local development (no cloud dependencies)
local = [
    "litellm",
    "aisuite[all]",
    "jsonschema>=4.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "flatmachines>=0.10.0",
]
# Alias for local
all = [
    "litellm",
    "aisuite[all]",
    "jsonschema>=4.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "flatmachines>=0.10.0",
]
# GCP deployment (local + Firestore)
gcp = [
    "litellm",
    "aisuite[all]",
    "jsonschema>=4.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "google-cloud-firestore>=2.11.0",
    "flatmachines>=0.10.0",
]

[project.urls]
Homepage = "https://github.com/memgrafter/flatagents"
Repository = "https://github.com/memgrafter/flatagents"
Issues = "https://github.com/memgrafter/flatagents/issues"

[tool.hatch.build.targets.wheel]
packages = ["flatagents"]
artifacts = [
    "flatagents/assets/*.d.ts",
    "flatagents/assets/*.json",
]
